+++ src/main/java/com/marketing/web/dtos/obligation/WritableObligation.java
@@ -5,14 +5,15 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 public class WritableObligation implements Serializable {
 
-    private Double debt;
+    private BigDecimal debt;
 
-    private Double receivable;
+    private BigDecimal receivable;
 
 }
