+++ src/main/java/com/marketing/web/dtos/product/ReadablePromotion.java
@@ -1,10 +1,12 @@
 package com.marketing.web.dtos.product;
 
+import com.marketing.web.enums.PromotionType;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 @Data
 @NoArgsConstructor
@@ -13,7 +15,7 @@ public class ReadablePromotion implements Serializable {
 
     private String promotionText;
 
-    private double discountValue;
+    private BigDecimal discountValue;
 
     private int discountUnit;
